Forgot to cast to float.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@3327 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ GuiFramework/Controller/GUIManager.cs
@@ -223,7 +223,7 @@ namespace Anomalous.GuiFramework
 
         private void ScreenLayoutManager_ScreenSizeChanging(int width, int height)
         {
-            this.CompactMode = width / height < .65f;
+            this.CompactMode = (float)width / height < .65f;
         }
 
         public event ScreenSizeChangedDelegate ScreenSizeChanged
